DN-4: Draft 1 of user accounts system, and re-organisation.
