+++ server/src/index.ts
@@ -54,6 +54,7 @@ app.get('/', async (req: Request, res: Response) => {
 });
 
 app.use('/', routes.auth);
+app.use('/', routes.test);
 // -----
 
 if (!env.BUILD_ENV_BOOL) {
