+++ server/src/routes/index.ts
@@ -1,1 +1,2 @@
-export { default as auth } from './auth/auth';
+export { default as auth } from './auth/index';
+export { default as test } from './test/protected';
