+++ server/src/utils/constants.ts
@@ -11,3 +11,9 @@ export function DATETIME_YYYY_MM_DD_HHMMSS(): string {
 }
 
 export const MAX_COMMENT_DEPTH: number = 5;
+
+export const SINGLE_HOUR: number = 60 * 60 * 1000;
+
+const hours: number = 3;
+export const TOKEN_VALIDITY_HOURS: number = 60 * 60 * 1000 * hours;
+export const TOKEN_SIZE: number = 64;
